requirements.txt
absl-py==2.1.0
arrow==1.3.0
astunparse==1.6.3
attrs==24.2.0
blinker
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.0
flatbuffers==24.3.25
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-api-python-client==2.48.0
google-auth==2.17.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
Google-Images-Search==1.4.7
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.11.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.2
keras==2.13.1
Markdown==3.4.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pyfiglet==0.8.post1
Pygments==2.15.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-resize-image==1.1.20
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.7.2
scipy==1.10.1
selenium==4.10.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.15
tensorboard==2.13.0
tensorflow-cpu==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
text-unidecode==1.3
typing_extensions==4.5.0
uritemplate==3.0.1
urllib3==2.0.4
websocket-client==1.5.2
WTForms==3.0.1
zipp==3.15.0
